requirements.txt
allure-pytest
allure-python-commons
async-generator
attrs
Babel
certifi
cffi
cryptography
h11
idna
iniconfig
outcome
packaging
pip
pluggy
py-moneyed
py
pycparser
pyOpenSSL
pyparsing
PySocks
pytest
pytz
selenium
setuptools
six
sniffio
sortedcontainers
splinter
stere
tomli
trio-websocket
trio
urllib3
wheel
wsproto
requests
Pillow
ImageHash